+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - improved prompt field with autocompletion
 - calculating cost including cached tokens for Gemini models
 - tool result containing image data is passed as image message to models
+- added duplicate tool call prevention in Agent
 
 ## [0.17.0]